+++ field/src/helpers.rs
@@ -1,0 +1,100 @@
+use std::collections::{HashMap, HashSet};
+
+use itertools::Itertools;
+
+use bempp_tree::{types::{morton::MortonKey, domain::Domain}};
+
+use crate::types::TransferVector;
+
+/// Algebraically defined list of unique M2L interactions, called 'transfer vectors', for 3D FMM.
+pub fn compute_transfer_vectors() -> Vec<TransferVector> {
+    let point = [0.5, 0.5, 0.5];
+    let domain = Domain {
+        origin: [0., 0., 0.],
+        diameter: [1., 1., 1.],
+    };
+
+    // Encode point in centre of domain
+    let key = MortonKey::from_point(&point, &domain, 3);
+
+    // Add neighbours, and their resp. siblings to v list.
+    let mut neighbours = key.neighbors();
+    let mut keys: Vec<MortonKey> = Vec::new();
+    keys.push(key);
+    keys.append(&mut neighbours);
+
+    for key in neighbours.iter() {
+        let mut siblings = key.siblings();
+        keys.append(&mut siblings);
+    }
+
+    // Keep only unique keys
+    let keys: Vec<&MortonKey> = keys.iter().unique().collect();
+
+    let mut transfer_vectors: Vec<usize> = Vec::new();
+    let mut targets: Vec<MortonKey> = Vec::new();
+    let mut sources: Vec<MortonKey> = Vec::new();
+
+    for key in keys.iter() {
+        // Dense v_list
+        let v_list = key
+            .parent()
+            .neighbors()
+            .iter()
+            .flat_map(|pn| pn.children())
+            .filter(|pnc| !key.is_adjacent(pnc))
+            .collect_vec();
+
+        // Find transfer vectors for everything in dense v list of each key
+        let tmp: Vec<usize> = v_list
+            .iter()
+            .map(|v| key.find_transfer_vector(v))
+            .collect_vec();
+
+        transfer_vectors.extend(&mut tmp.iter().cloned());
+        sources.extend(&mut v_list.iter().cloned());
+
+        let tmp_targets = vec![**key; tmp.len()];
+        targets.extend(&mut tmp_targets.iter().cloned());
+    }
+
+    let mut unique_transfer_vectors = Vec::new();
+    let mut unique_indices = HashSet::new();
+
+    for (i, vec) in transfer_vectors.iter().enumerate() {
+        if !unique_transfer_vectors.contains(vec) {
+            unique_transfer_vectors.push(*vec);
+            unique_indices.insert(i);
+        }
+    }
+
+    let unique_sources: Vec<MortonKey> = sources
+        .iter()
+        .enumerate()
+        .filter(|(i, _)| unique_indices.contains(i))
+        .map(|(_, x)| *x)
+        .collect_vec();
+
+    let unique_targets: Vec<MortonKey> = targets
+        .iter()
+        .enumerate()
+        .filter(|(i, _)| unique_indices.contains(i))
+        .map(|(_, x)| *x)
+        .collect_vec();
+
+    let mut result = Vec::<TransferVector>::new();
+
+    for ((t, s), v) in unique_targets
+        .into_iter()
+        .zip(unique_sources)
+        .zip(unique_transfer_vectors)
+    {
+        result.push(TransferVector {
+            vector: v,
+            target: t,
+            source: s,
+        })
+    }
+
+    result
+}
